Move generateCraftedItem to gm.cmd
